+++ src/lsr_wiping.c
@@ -378,9 +378,9 @@ __lsr_fill_buffer (
 			do
 			{
 # if (!defined __STRICT_ANSI__) && (defined HAVE_RANDOM)
-				i = (size_t) ((size_t)random () % npat);
+				i = (size_t)random () % npat;
 # else
-				i = (size_t) ((size_t)rand () % npat);
+				i = (size_t)rand () % npat;
 # endif
 			}
 			while ( (selected[i] == 1) && (__lsr_sig_recvd () == 0) );
